changed yscale to a double
as an int, this has issues if the number of a y point is larger than the number of pixels available to display (ends up forcing scale to 0 and thus all points end up at 0)

+++ OVGraphView/OVGraphView/OVGraphPlotView.h
@@ -11,7 +11,7 @@
 #import "OVGraphYIndicatorView.h"
 @class OVPlotScrollView;
 @interface OVGraphPlotView : UIView <UIScrollViewDelegate>{
-    int yscale;
+    double yscale;
     int spacebetweenpoints;
     UIColor *opaquecolor;
     UIColor *lightercolor;
